% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal-apis.R
\name{.inheritParams}
\alias{.inheritParams}
\title{Api internal functions}
\arguments{
\item{competition_code}{The code of the competition.\cr
Admitted values are \code{E} for Euroleague and \code{U} for Eurocup.}

\item{team_code}{The code of the team.\cr
Examples are \code{ASV}, \code{MAD}, ...}

\item{game_code}{The code of the game as obtained from \code{\link[=getCompetitionGames]{getCompetitionGames()}}.}

\item{season_code}{The code of the season as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}

\item{round}{The value of the round as obtained from \code{\link[=getCompetitionRounds]{getCompetitionRounds()}}.}

\item{phase_type}{The code of competition phase.\cr
Admitted values are \code{RS} for regular season, \code{PO} for playoffs and \code{FF} for final four.
Default is \code{All} for all.}

\item{subset}{A particular subset of games.\cr
Admitted values are \code{HomeGames}, \code{AwayGames}, \code{GamesWon}, \code{GamesLost},
\code{ResultsIn5Points} (for games resulted in +/-5 points) and \code{All}.
Default is \code{All}.}

\item{statistic_mode}{Type of aggregation of statistics.\cr
Admitted values are \code{perGame}, \code{perMinute} and \code{accumulated}.}
}
\description{
Primitive Api wrapper functions are not exported since they only allow for one
call at a time.
Instead, an equivalent functions is exposed allowing multiple
parameters, making it easier to retrieve data for users.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
