% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{euroleague_pbp}
\alias{euroleague_pbp}
\title{Get Euroleague play-by-play data}
\usage{
euroleague_pbp(season, verbose = FALSE)
}
\arguments{
\item{season}{either a single integer or range of seasons}

\item{verbose}{set to TRUE if you want season progress alert}
}
\value{
returns a data frame of play-by-play data
}
\description{
Returns a data frame of all available play by play
data for any season(s) of Euroleague matches played
since 2010-11 season, including play types, players
involved and outcomes.
}
\examples{
try({
pbp_22 <- euroleague_pbp(season = 2022)
pbp_21to22 <- euroleague_pbp(season = c(2021:2022))
})
}
