% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{euroleague_player_box}
\alias{euroleague_player_box}
\title{Get Euroleague player box score data}
\usage{
euroleague_player_box(season, verbose = FALSE)
}
\arguments{
\item{season}{either a single integer or range of seasons}

\item{verbose}{set to TRUE if you want season progress alert}
}
\value{
returns a dataframe of player box score data
}
\description{
Returns a data frame of all available player box scores
data for Euroleague matches played since 2010-11
}
\examples{
try({
player_box <- euroleague_player_box(season = 2022)
player_box_21to22 <- euroleague_player_box(season = c(2021:2022))
})
}
