% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{euroleague_results}
\alias{euroleague_results}
\title{Get match results for all Euroleague matches}
\usage{
euroleague_results()
}
\value{
returns a data frame of match results
}
\description{
Returns a data frame of all results of all Eoroleague matches
played since the 2000-2001 season
}
\examples{
try({
match_results <- euroleague_results()
})
}
