% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{euroleague_shots}
\alias{euroleague_shots}
\title{Get Euroleague shot data}
\usage{
euroleague_shots(season, verbose = FALSE)
}
\arguments{
\item{season}{either a single integer or range of seasons}

\item{verbose}{set to TRUE if you want season progress alert}
}
\value{
returns a data frame of shot location data
}
\description{
Returns a data frame of all available shot data
for any season(s) of Euroleague matches played
since 2010-11 season, including shot types,
shot locations and players.
}
\examples{
try({
shots_22 <- euroleague_shots(season = 2022)
shots_20to22 <- euroleague_shots(season = c(2020:2022))
})
}
