% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{euroleague_team_box}
\alias{euroleague_team_box}
\title{Get Euroleague team box score data}
\usage{
euroleague_team_box(season, verbose = FALSE)
}
\arguments{
\item{season}{either a single integer or range of seasons}

\item{verbose}{set to TRUE if you want season progress alert}
}
\value{
returns a dataframe of team box score data
}
\description{
Returns a data frame of all available team box scores
data for Euroleague matches played since 2010-11
}
\examples{
try({
team_box <- euroleague_team_box(season = 2022)
team_box_21to22 <- euroleague_team_box(season = c(2021:2022))
})
}
