% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompetitionHistory.R
\name{getCompetitionHistory}
\alias{getCompetitionHistory}
\title{Get competition history}
\usage{
getCompetitionHistory(competition_code)
}
\arguments{
\item{competition_code}{One or more competition codes.\cr
Admitted values are \code{E} for Euroleague and \code{U} for Eurocup.}
}
\value{
Returns a summary tibble of history for chosen competitions
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getCompetitionHistory(competition_code = c("E", "U"))

}
