% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-competition-phases.R
\name{getCompetitionPhases}
\alias{getCompetitionPhases}
\title{Get competition phases}
\usage{
getCompetitionPhases(season_code)
}
\arguments{
\item{season_code}{One or more season codes as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}
}
\value{
Returns a summary tibble of phases for chosen competitions
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getCompetitionPhases(season_code = c("E2023", "U2023"))

}
\keyword{competitionMetadata}
