% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-competition-rounds.R
\name{getCompetitionRounds}
\alias{getCompetitionRounds}
\title{Get competition rounds}
\usage{
getCompetitionRounds(season_code)
}
\arguments{
\item{season_code}{The code of the season as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}
}
\value{
Returns a summary tibble of rounds for chosen competitions
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getCompetitionRounds(season_code = c("E2023", "E2022"))

}
