% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-game-evolution.R
\name{getGameEvolution}
\alias{getGameEvolution}
\title{Get game evolution}
\usage{
getGameEvolution(game_code, season_code)
}
\arguments{
\item{game_code}{The code of the game. An integer value}

\item{season_code}{The code of the season (examples are "E2023" for Euroleague
or "U2023" for Eurocup 2023)}
}
\value{
Returns a list of two elements for the chosen games and seasons
\itemize{
\item \emph{EvolutionSummary}. Overall information about minimum and maximum difference of scores
between teams
\item \emph{Evolution}. Minute by minute points of each team
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/game-center/2023-24/crvena-zvezda-meridianbet-belgrade-ldlc-asvel-villeurbanne/E2023/1/#boxscore}{BoxScore}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getGameEvolution(season_code = c("E2023", "U2023"), game_code = 1)

}
