% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-game-playbyplay.R
\name{getGamePlayByPlay}
\alias{getGamePlayByPlay}
\title{Get game play-by-play}
\usage{
getGamePlayByPlay(game_code, season_code)
}
\arguments{
\item{game_code}{The code of the game as obtained from \code{\link[=getCompetitionGames]{getCompetitionGames()}}.}

\item{season_code}{The code of the season as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}
}
\value{
Returns a list of two elements for the chosen games and seasons
\itemize{
\item \strong{PlayByPlaySummary}. Overall information about the games, teams involved and status (live or not)
\item \strong{PlayByPlay}. Detailed information about the games, particularly \strong{NumberOfPlay} and \strong{TypeOfPlay}
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/game-center/2023-24/crvena-zvezda-meridianbet-belgrade-ldlc-asvel-villeurbanne/E2023/1/#play-by-play}{PlayByPlay}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getGamePlayByPlay(season_code = c("E2023", "U2023"), game_code = 1)$PlayByPlay

}
