% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-game-players.R
\name{getGamePlayers}
\alias{getGamePlayers}
\title{Get game players}
\usage{
getGamePlayers(game_code, team_code, season_code)
}
\arguments{
\item{game_code}{The code of the game. An integer value}

\item{team_code}{The code of the team (examples are "ASV", "MAD", ...)}

\item{season_code}{The code of the season (examples are "E2023" for Euroleague
or "U2023" for Eurocup 2023)}
}
\value{
Returns a tibble of players with related information for the given game
and season
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getGamePlayers(season_code = c("E2023", "U2023"), team_code = "ASV", game_code = 1)

}
