% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-player-advanced.R
\name{getPlayerAdvanced}
\alias{getPlayerAdvanced}
\title{Get player advanced statistics}
\usage{
getPlayerAdvanced(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)
}
\arguments{
\item{season_code}{One or more season codes as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}

\item{statistic_mode}{One or more aggregation modes of statistics.\cr
Admitted values are \code{perGame}, \code{perMinute} and \code{accumulated}.}
}
\value{
Returns a summary tibble of advanced players statistics for chosen seasons
\itemize{
\item \strong{eFG\%} (Effective field goal \%) - Combined two- and three-point shooting effectiveness
\item \strong{TS\%} (True shooting \%) - Percentage of points vs. points attempted
\item \strong{OREB\%} (Offensive rebound \%) - Estimated \% of available offensive rebounds obtained while on court
\item \strong{DREB\%} (Defensive rebound \%) - Estimated \% of available defensive rebounds obtained while on court
\item \strong{REB\%} (Rebound \%) - Estimated \% of available rebounds obtained while on court
\item \strong{AST/TO} (Assist to turnover ratio) - Ratio of assists made to turnovers committed
\item \strong{AST-R} (Assist ratio) - Estimated \% of assists per player's offensive possessions
\item \strong{TO-R} (Turnover ratio) - Estimated \% of turnovers per player's offensive possessions
\item \strong{2PTA-R} (Two-point attempts ratio) - Estimated \% of two-point attempts per player's offensive possessions
\item \strong{3PTA-R} (Three-point attempts ratio) - Estimated \% of three-point attempts per player's offensive possessions
\item \strong{FT-RT} (Free Throw rate) - Measure of free throw attempts vs. field goal attempts
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/stats/expanded/?size=1000&viewType=advanced&seasonCode=E2023&statisticMode=perGame&seasonMode=Single&sortDirection=descending&statistic=blocksFavour}{Stats}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getPlayerStats(season_code = "E2023", statistic_mode = "perGame")

}
