% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-player-points.R
\name{getPlayerPoints}
\alias{getPlayerPoints}
\title{Get player points statistics}
\usage{
getPlayerPoints(season_code)
}
\arguments{
\item{season_code}{The code of the season as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}
}
\value{
Returns a summary tibble of points players statistics for chosen seasons
\itemize{
\item \strong{2PTA-S} (Two-point attempts share) - Player's share of team's total two-point attempts
\item \strong{3PTA-S} (Three-point attempts share) - Player's share of team's total three-point attempts
\item \strong{FTA-S} (Free throw attempts share) - Player's share of team's total free throw attempts
\item \strong{2PTM-S} (Two-pointers made share) - Player's share of team's total two-pointers made
\item \strong{3PTM-S} (Three-pointers made share) - Player's share of team's total three-pointers made
\item \strong{FTM-S} (Free throws made share) - Player's share of team's total free throws made
\item \strong{2PT-RT} (Two-Point Rate) - \% of a player's field goal attempts that are two-pointers
\item \strong{3PT-RT} (Three-Point Rate) - \% of field goal attempts that are three-pointers
\item \strong{\%2PT} (\% of points from two-pointers) - \% of points from two-point shots made
\item \strong{\%3PT} (\% of points from three-pointers) - \% of points from three-point shots made
\item \strong{\%FT} (\% of points from free throws) - \% of points from free throws made
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/stats/expanded/?size=1000&viewType=scoring&seasonCode=E2023&statisticMode=perGame&seasonMode=Single&sortDirection=descending&statistic=fieldGoals2Percent}{Stats}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getPlayerPoints(season_code = "E2023")

}
