% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-team-lead-stats.R
\name{getTeamLeadStats}
\alias{getTeamLeadStats}
\title{Get team lead statistics}
\usage{
getTeamLeadStats(season_code, phase_type = "All", subset = "All")
}
\arguments{
\item{season_code}{One or more season codes as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}

\item{phase_type}{One or more phase type codes as obtained from \code{\link[=getCompetitionPhases]{getCompetitionPhases()}}.\cr
Admitted values are \code{RS} for regular season, \code{PO} for playoffs and \code{FF} for final four.
Default is \code{All} for all.}

\item{subset}{One or more game subsets.\cr
Admitted values are \code{HomeGames}, \code{AwayGames}, \code{GamesWon}, \code{GamesLost},
\code{ResultsIn5Points} (for games resulted in +/-5 points) and \code{All}.
Default is \code{All}.}
}
\value{
Returns a list of elements for the chosen seasons, phase_type and subset.
\itemize{
\item \strong{TeamAccumulated}. Total sum of statistics of all teams team
\item \strong{TeamAveragePerGame}. Average per game of statistics of all teams
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/stats/key-stats-teams/?size=200&misc=HomeGames&seasonCode=E2023&category=FreeThrowsAttempted&seasonMode=Single&sortDirection=descending&aggregate=accumulated}{TeamLead}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
TeamLeadStats = getTeamLeadStats(season_code = c("E2022", "E2023"), phase_type = "RS")

TeamLeadStats$TeamAccumulated |> head(5)

TeamLeadStats$TeamAveragePerGame |> head(5)

}
