% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-team-people.R
\name{getTeamPeople}
\alias{getTeamPeople}
\title{Get team people}
\usage{
getTeamPeople(team_code, season_code)
}
\arguments{
\item{team_code}{The code of the team.\cr
Examples are \code{ASV}, \code{MAD}, ...}

\item{season_code}{The code of the season as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}
}
\value{
Returns summary information about team people by season

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/teams/ldlc-asvel-villeurbanne/roster/asv/?season=2023-24}{Team}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getTeamPeople(team_code = "ASV", season_code = c("E2023", "E2022"))

}
