% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-team-stats.R
\name{getTeamStats}
\alias{getTeamStats}
\title{Get team stats}
\usage{
getTeamStats(team_code, season_code, phase_type = "")
}
\arguments{
\item{team_code}{The code of the team (examples are "ASV", "MAD", ...)}

\item{season_code}{The code of the season (examples are "E2023" for Euroleague
or "U2023" for Eurocup 2023)}

\item{phase_type}{The code of competition phase ("RS" for regular season, "PO"
for playoffs and "FF" for final four). Default is "" for all.}
}
\value{
Returns a list of elements for the chosen teams and seasons
\itemize{
\item \emph{PlayerAccumulated}. Total sum of statistics by player
\item \emph{PlayerAveragePerGame}. Average per game of statistics by player
\item \emph{PlayerAveragePer40}. Average per 40 minutes of statistics by player
\item \emph{TeamAccumulated}. Total sum of statistics of team
\item \emph{TeamAveragePerGame}. Average per game of statistics of teams
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/teams/ldlc-asvel-villeurbanne/statistics/asv/?season=2023-24&phase=All\%20phases}{Team}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getTeamStats(team_code = "ASV", season_code = c("E2023", "E2022"), phasetype = "RS")

}
