% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompetitionStandings.R
\name{getCompetitionStandings}
\alias{getCompetitionStandings}
\title{Get competition traditional standings}
\usage{
getCompetitionStandings(season_code, round)
}
\arguments{
\item{season_code}{One or more season codes as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}

\item{round}{One or more round codes as obtained from \code{\link[=getCompetitionRounds]{getCompetitionRounds()}}.}
}
\value{
Returns a summary tibble of standings for chosen competitions and rounds

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/standings/}{Competition}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getCompetitionStandings(season_code = c("E2023", "E2022"), round = 1)

}
