% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-competition-standings.R
\name{getCompetitionStandings}
\alias{getCompetitionStandings}
\title{Get competition standings}
\usage{
getCompetitionStandings(season_code, round)
}
\arguments{
\item{season_code}{The code of the season (examples are \code{E2023} for Euroleague
or \code{U2023} for Eurocup 2023)}

\item{round}{The value of the round. An integer value}
}
\value{
Returns a summary tibble of standings for choseN competition and round

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/standings/}{Competition}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getCompetitionStandings(season_code = c("E2023", "E2022"), round = 1)

}
