% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-game-box-score.R
\name{getGameBoxScore}
\alias{getGameBoxScore}
\title{Get game box-score}
\usage{
getGameBoxScore(game_code, season_code)
}
\arguments{
\item{game_code}{The code of the game as obtained from \code{\link[=getCompetitionGames]{getCompetitionGames()}}.}

\item{season_code}{The code of the season as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}
}
\value{
Returns a list of elements for the chosen games and seasons
\itemize{
\item \strong{Team}. Name of the teams
\item \strong{Coach}. Name of the coaches
\item \strong{EndOfQuarter}. Team accumulated points by quarter
\item \strong{ByQuarter}. Team points for each quarter
\item \strong{PlayerStats}. Statistics for each player in the game
\item \strong{TeamStats}. Aggregated statistics for each team in the game
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/game-center/2023-24/crvena-zvezda-meridianbet-belgrade-ldlc-asvel-villeurbanne/E2023/1/#boxscore}{BoxScore}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getGameBoxScore(season_code = c("E2023", "U2023"), game_code = 1)

}
