% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGameHeader.R
\name{getGameHeader}
\alias{getGameHeader}
\title{Get game header}
\usage{
getGameHeader(game_code, season_code = "E2023")
}
\arguments{
\item{game_code}{The code of the game}

\item{season_code}{The code of the season}
}
\description{
Description...
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getGameHeader(game_code = 1, season_code = "E2023")

}
