% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-game-players.R
\name{getGamePlayers}
\alias{getGamePlayers}
\title{Get game players}
\usage{
getGamePlayers(game_code, team_code, season_code)
}
\arguments{
\item{game_code}{One or more game codes as obtained from \code{\link[=getCompetitionGames]{getCompetitionGames()}}.}

\item{team_code}{One or more team codes as obtained from \code{\link[=getCompetitionTeams]{getCompetitionTeams()}}.\cr
Examples are \code{ASV}, \code{MAD}, ...}

\item{season_code}{One or more season codes as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}
}
\value{
Returns a tibble of players with related information for the chosen games
and seasons
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getGamePlayers(season_code = c("E2023", "U2023"), team_code = "ASV", game_code = 1)

}
