% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-game-points.R
\name{getGamePoints}
\alias{getGamePoints}
\title{Get game points}
\usage{
getGamePoints(game_code, season_code)
}
\arguments{
\item{game_code}{The code of the game. An integer value.}

\item{season_code}{The code of the season (examples are "E2023" for Euroleague
or "U2023" for Eurocup 2023)}
}
\value{
Returns scoring information of each player for the chosen games and seasons (subset of play-by-play data).
In particular:
\itemize{
\item NumAnot. Reference id of the action (useful for join with retult of getPlayByPlay)
\item CoordX and CoordY. Spatial coordinates of the shot
\item Zone. Area of the court of the shot
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/game-center/2023-24/crvena-zvezda-meridianbet-belgrade-ldlc-asvel-villeurbanne/E2023/1/#play-by-play}{PlayByPlay}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getGamePoints(season_code = c("E2023", "U2023"), game_code = 1)

}
