% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-player-stats.R
\name{getPlayerStats}
\alias{getPlayerStats}
\title{Get player statistics}
\usage{
getPlayerStats(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)
}
\arguments{
\item{season_code}{The code of the season (examples are \code{E2023} for Euroleague
or \code{U2023} for Eurocup 2023)}

\item{statistic_mode}{Type of aggregation of statistics. Available values are
\code{perGame}, \code{perMinute} or \code{accumulated}}
}
\value{
Returns a summary tibble of players statistics for chosen competition

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/stats/expanded/?size=1000&viewType=traditional&seasonCode=E2023&statisticMode=perGame&seasonMode=Single&sortDirection=descending&statistic=blocksFavour}{Competition}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getPlayerStats(season_code = "E2023", statistic_mode = "perGame")

}
