% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-player-stats.R
\name{getPlayerStats}
\alias{getPlayerStats}
\title{Get player statistics}
\usage{
getPlayerStats(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)
}
\arguments{
\item{season_code}{One or more season codes as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}

\item{statistic_mode}{One or more aggregation modes of statistics.\cr
Admitted values are \code{perGame}, \code{perMinute} and \code{accumulated}.}
}
\value{
Returns a summary tibble of players statistics for chosen seasons
\itemize{
\item \strong{PTS} (Points scored)
\item \strong{2PM} (Two-pointers made)
\item \strong{2PA} (Two-pointers attempted)
\item \strong{2P\%} (Two-point \%)
\item \strong{3PM} (Three-pointers made)
\item \strong{3PA} (Three-pointers attempted)
\item \strong{3P\%} (Three-point \%)
\item \strong{FTM} (Free throws made)
\item \strong{FTA} (Free throws attempted)
\item \strong{FT\%} (Free-throw \%)
\item \strong{OR} (Offensive rebounds)
\item \strong{DR} (Defensive rebounds)
\item \strong{TR} (Total rebounds)
\item \strong{AST} (Assists)
\item \strong{STL} (Steals)
\item \strong{TO} (Turnovers)
\item \strong{BLK} (Blocks)
\item \strong{BLKA} (Blocks against)
\item \strong{FC} (Personal fouls committed)
\item \strong{FD} (Personal fouls drawn)
\item \strong{PIR} (Performance Index Rating)
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/stats/expanded/?size=1000&viewType=traditional&seasonCode=E2023&statisticMode=perGame&seasonMode=Single&sortDirection=descending&statistic=blocksFavour}{Stats}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getPlayerStats(season_code = "E2023", statistic_mode = "perGame")

}
