% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-player-advanced.R, R/get-player-stats.R
\name{getPlayerStats}
\alias{getPlayerStats}
\title{Get player advanced statistics}
\usage{
getPlayerStats(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)

getPlayerStats(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)
}
\arguments{
\item{season_code}{The code of the season (examples are \code{E2023} for Euroleague
or \code{U2023} for Eurocup 2023)}

\item{statistic_mode}{Type of aggregation of statistics. Available values are
\code{perGame}, \code{perMinute} or \code{accumulated}}
}
\value{
Returns a summary tibble of advanced players statistics for chosen seasons
\itemize{
\item eFG\% (Effective field goal \%) - Combined two- and three-point shooting effectiveness
\item TS\% (True shooting \%) - Percentage of points vs. points attempted
\item OREB\% (Offensive rebound \%) - Estimated \% of available offensive rebounds obtained while on court
\item DREB\% (Defensive rebound \%) - Estimated \% of available defensive rebounds obtained while on court
\item REB\% (Rebound \%) - Estimated \% of available rebounds obtained while on court
\item AST/TO (Assist to turnover ratio) - Ratio of assists made to turnovers committed
\item AST-R (Assist ratio) - Estimated \% of assists per player's offensive possessions
\item TO-R (Turnover ratio) - Estimated \% of turnovers per player's offensive possessions
\item 2PTA-R (Two-point attempts ratio) - Estimated \% of two-point attempts per player's offensive possessions
\item 3PTA-R (Three-point attempts ratio) - Estimated \% of three-point attempts per player's offensive possessions
\item FT-RT (Free Throw rate) - Measure of free throw attempts vs. field goal attempts
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/stats/expanded/?size=1000&viewType=advanced&seasonCode=E2023&statisticMode=perGame&seasonMode=Single&sortDirection=descending&statistic=blocksFavour}{Stats}

Returns a summary tibble of players statistics for chosen seasons
\itemize{
\item PTS (Points scored)
\item 2PM (Two-pointers made)
\item 2PA (Two-pointers attempted)
\item 2P\% (Two-point \%)
\item 3PM (Three-pointers made)
\item 3PA (Three-pointers attempted)
\item 3P\% (Three-point \%)
\item FTM (Free throws made)
\item FTA (Free throws attempted)
\item FT\% (Free-throw \%)
\item OR (Offensive rebounds)
\item DR (Defensive rebounds)
\item TR (Total rebounds)
\item AST (Assists)
\item STL (Steals)
\item TO (Turnovers)
\item BLK (Blocks)
\item BLKA (Blocks against)
\item FC (Personal fouls committed)
\item FD (Personal fouls drawn)
\item PIR (Performance Index Rating)
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/stats/expanded/?size=1000&viewType=traditional&seasonCode=E2023&statisticMode=perGame&seasonMode=Single&sortDirection=descending&statistic=blocksFavour}{Stats}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getPlayerStats(season_code = "E2023", statistic_mode = "perGame")

getPlayerStats(season_code = "E2023", statistic_mode = "perGame")

}
