% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-team-games.R
\name{getTeamGames}
\alias{getTeamGames}
\title{Get team games}
\usage{
getTeamGames(team_code, season_code)
}
\arguments{
\item{team_code}{The code of the team (examples are "ASV", "MAD", ...)}

\item{season_code}{The code of the season (examples are "E2023" for Euroleague
or "U2023" for Eurocup 2023)}
}
\value{
Returns summary information about team games by season

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/teams/ldlc-asvel-villeurbanne/games/asv/?season=2023-24}{Team}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
getTeamGames(team_code = "ASV", season_code = c("E2023", "E2022"))

}
