% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-team-stats.R
\name{getTeamStats}
\alias{getTeamStats}
\title{Get team statistics}
\usage{
getTeamStats(team_code, season_code, phase_type = "All")
}
\arguments{
\item{team_code}{One or more team codes as obtained from \code{\link[=getCompetitionTeams]{getCompetitionTeams()}}.\cr
Examples are \code{ASV}, \code{MAD}, ...}

\item{season_code}{One or more season codes as obtained from \code{\link[=getCompetitionHistory]{getCompetitionHistory()}}.\cr
Examples are \code{E2023} for Euroleague or \code{U2023} for Eurocup 2023.}

\item{phase_type}{One or more phase type codes as obtained from \code{\link[=getCompetitionPhases]{getCompetitionPhases()}}.\cr
Admitted values are \code{RS} for regular season, \code{PO} for playoffs and \code{FF} for final four.
Default is \code{All} for all.}
}
\value{
Returns a list of elements for the chosen teams and seasons
\itemize{
\item \strong{PlayerAccumulated}. Total sum of statistics by player
\item \strong{PlayerAveragePerGame}. Average per game of statistics by player
\item \strong{PlayerAveragePer40}. Average per 40 minutes of statistics by player
\item \strong{TeamAccumulated}. Total sum of statistics of team
\item \strong{TeamAveragePerGame}. Average per game of statistics of teams
}

Reference webpage: \href{https://www.euroleaguebasketball.net/euroleague/teams/ldlc-asvel-villeurbanne/statistics/asv/?season=2023-24&phase=All\%20phases}{Team}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
TeamStats = getTeamStats(team_code = "ASV", season_code = c("E2023", "E2022"), phase_type = "RS")
TeamStats$PlayerAveragePerGame

}
